02-23 커뮤니티 댓글 삭제, 커뮤니티 삭제 진행중

+++ src/main/java/com/green/gogiro/community/model/CommunityCommentInsDto.java
@@ -11,7 +11,7 @@ import lombok.Data;
 @Data
 public class CommunityCommentInsDto {
     @Schema(title = "커뮤니티pk")
-    private int iboard;
+    private Long iboard;
 
     @JsonIgnore
     private int iuser;
